crash_tracker: remove old crash files
The broker will retain the 50 most recent crash files and delete all
older crashes on startup. This is to ensure that crash report files do
not fill up the disk.

Note that we remove files before creating them to ensure cleanup
succeeds before we create new files. This means that technically we will
have up to 51 crash reports on disk (50 kept + 1 created). This is
acceptable since 50 is a loose limit.