ducktape: Add defaults to ScaleParameters
Add default values for partition limit related values to ScaleParameters
so that tests don't need to duplicate them. These need to be kept
in sync with configuration.cc and most tests should automatically pick
up the new values when they change.

Use the defaults as default values for the constructor parameters, so
that most tests can use a simpler constructor and get the right values
which follow updates.