bazel: fix use custom junit report writer
The boost built in one has a BUNCH of ASAN violation and memory leaks.
In generally it seems pretty sketchy, so just write our own that makes
Bazel happy. It really only needs the list of test cases and if it
failed or not. It doesn't report on more information than that, and the
output here is similar to GoogleTest.