crypto/tests: do not use RP_FIXTURE_ENV for ossl_context_service tests
`ossl_context_service` only works if we have one seastar shard per
thread. RP_FIXTURE_ENV env variable disables the service to make it safe
to run multiple seastar shards in a single thread. A lot of tests need
this, so it is set by default. However, to test the service itself we
need to enable it.