pandaproxy/sr: Relax restrictions for json schema
Previous implementation was too strict with "id"/"$id" keywords.
"id" is a restricted keyword under draft-04 and "$id" is a
restricted keyword under all the others. There is nothing preventing
a non-draft-04 to define "id" and vice versa.

We should only consider the appropriate id keyword based on the active
schema draft. Not look for both.
Also removed a test case that is not valid under the relaxed conditions.