vassert: Created assert_log_holder
Created a structure that will be used to hold a callback function that
will be called when an assert is fired.  The desire will be for the
crash tracker module to register a callback that will trigger creation
of a crash tracker file containing the vassert message and stack trace.

A call back was used rather than directly calling into the crash tracker
from assert in order to not make crash tracker a global depenency, as
vassert is today.

Signed-off-by: Michael Boquard <michael@redpanda.com>