ct: Fix unit-tests in cloud topics
Some tests which relying on manual clock do not work correctly with
debug seastar build without the explicit sleep.

Some tests are relying on the utility that advances the manual clock's
time forward and then invokes 'seastar::yield' in a loop. With the debug
mode Seastar scheduler this doesn't work. If the task is scheduled to
run in the background (not awaited) and then the utility is invoked the
task never gets completed and the utility reports 'timeout' error.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>