kafka: move bad ts metric to kafka server
Kafka bad timestamp metric is associated with generic net::server base
class probe, which doesn't make sense since it's a kafka server concept
(so, for example this metric will appear on the internal rpc server,
which also derives from this class, though the metric doesn't sense
there and cannot be incremented there). Move it to kafka server probe
instead. The full name remains the same so there will be no change for
the end-user.