`storage`: add `disk_log_impl::reset_dirty_and_closed_bytes()`
Acts as a safety hatch during calls to `disk_log_impl::dirty_ratio()`
against negative, bogus values in `_dirty_segment_bytes` and
`_closed_segment_bytes` by performing an `O(n)` summation of the proper
dirty and closed bytes values over each segment in the log.

It is expected that `disk_log_impl::dirty_ratio()` is the public
function used to probe these values by external services, and that is why
the safety hatch is triggered there and not in one of the `add/subtract`
functions. This approach also has the added benefit of allowing the
`dassert()`s in those functions to still trigger on negative values, and
help identify any potential remaining bugs in the bytes bookkeeping.