crash_tracker: record the app version
Currently we can use the heartbeat dataset to join the cluster uuid of a
telemetry-reported crash to figure out which version of redpanda crashes
and therefore which redpanda binary we need to use to decode the crash
stacktrace. However, for crashes the happen around the time when a
cluster is upgraded, it is more reliable to record the redpanda build
version directly into the crash report. This will also be included in
the telemetry data.