dl/translation: add jitter to partition_translator
The previous implementation of the partition_translator had a jitter
between retries of translation. This went away once we moved to
scheduler-based translation, but led to situations where failures in
translation would get retried immediately, which could overwhelm the
catalog if the failure were caused by a catalog issue.

This adds a temporary fix to add jitter to the partition translator's
loop on early exit. A more complete solution may be to build backoffs
into the scheduler upon translation failure.

Along the way, this updates a couple of error cases to continue to the
next iteration early, which seemed like an oversight.