schema: make get API used by Iceberg more robust
We were previously getting schemas without explicitly syncing (listing
offsets and consuming _schemas until the current HWM) with the
underlying schema store. This meant that Iceberg translation could see a
stale result that misled it into thinking it was missing a schema.

This commit updates the API that Iceberg uses to sync if it detects that
the requested schema is missing. I opted to do this conditionally
because the sync seems somewhat expensive.

Without this change, datalake/cluster_restore_test.py was very flaky,
even when run locally (creating DLQ tables, not creating main tables).
With this change, it at least passes locally.