`archival`: check `first_addressable_segment` in `retention_calculator`
We were dereferencing the iterator returned from
`manifest.first_addressable_segment()` without first checking if this
iterator was valid or not.

In the case that `_start_offset == model::offset{}` in the
`partition_manifest`, this iterator will be `end()` and we will be
dereferencing an invalid iterator.

Fix this by adding a check that `first_seg != manifest.end()`.