cloud_storage: Update process_anomalies method
The current implementation has a bug. If the manifest is empty
the filtering of false positives doesn't work for missing segments.
This commit fixes the problem by filtering out all missing segments
if the manifest is empty or if all segments in the manifest are below
the start offset (which is the same thing).

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>