r/buffered_protocol: fixed potential uaf error in append_entries_queue
When processing append_entries_request in append entries queue it may
happen that after the semaphore returned but before the subsequent
future is scheduled the queue has been deleted. This might lead to UAF
error.

Holding a gate when processing append entries requests prevent the queue
from being disposed while waiting for the next request to be dispatched.

Signed-off-by: Michał Maślanka <michal@redpanda.com>