rpk: Prompt for positive confirmation
in case if the user is disabling 'redpanda.remote.write'. This is a
potentially dangerous operation that can lead to data loss. The prompt
suggest an alternaive for that (safe pause).

The '--no-confirm' flag that allows the user to bypass the prompt is
added. The flag is supposed to be used in scripts and CI.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>