datalake: use latest subject iceberg mode
Instead of a specific protobuf one, we'll just have a generic schema
mode for latest schema within a subject.

By default we'll use the topic naming strategy to determine the subject
and we'll assume the first protobuf message in a file. Otherwise it's
possible for a user to override these values using inline key/value
pairs to the topic config.