archival: Swap segment collector implementation
Previously, the search of segments that can be uploaded to the cloud
storage was implemented inside the archival_policy. For non-compacted
segments the archival_policy used its internal implementation but for
compacted reuploads it was using segment_collector.

This commit introduces new segment serach logic that uses
segment_collector so both compacted and non-compacted uploads are going
through the collector now.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>