cluster: enable shard_placement_table persistence if the feature is active
It is possible that the node_local_core_assignment feature flag gets
activated on a node after the startup, but the prepare phase (and the
associated migration) is skipped. This can happen if the node re-joins
with the same id after its disk was erased and node UUID overriden.

Enable persistence in this case as well, thereby bringing local kvstore
state in sync with the global feature table state.

Fixes https://redpandadata.atlassian.net/browse/CORE-9110