archival: archival retention calculator Kafka offset pin
Similar to the pin for spillover retention, this adds a mechanism to pin
archival STM retention based on on a given Kafka offset. Once Iceberg is
able to read from tiered storage, this will allow us to preserve the
semantics that regardless of retention policy, Kafka records are
guaranteed to land in Iceberg.