bazel: fix micro benchmarks
Do this by leveraging runfiles instead of environment variables, which
break when you switch directories. Care must be taken to keep
compatibility between CMake and Bazel, so wasm need to fall back to the
current directory for Wasm binaries, and OpenSSL tests are quite tricky
to get right with all the subtle usages of environment variables.

Also switch bench wrapper to python because the bash is becoming
non-trivial and we need to resolve some environment variables for
sanitizer configuration options.