crypto: simplify crypto_bench
Instead of really trying hard to keep the same instance of the
ossl_context_service during the entire test run, just clean it up and
shut it down for each instance of the test. This fixes a crash that I
don't fully understand where threads are not joined - which I assume
means that there is a case when the global variable is destructed but
at_exit isn't called for some reason (an exception thrown?). Either way
this prevents that.