tests/upgrade: added test verifying total bytes read during recovery
Added test verifying if there is an increase in the total bytes that are
read during the first startup of Redpanda after upgrade. The increased
read size may cause high disk and cpu usage in cluster with lots of data
retained locally.

The test makes sure that the total amount of bytes read after upgrade is
not increased more than 10%.

Signed-off-by: Michał Maślanka <michal@redpanda.com>