archival: Fix archival STM snapshot installation
The code that installs the archival STM snapshot during the recovery
uses the wrong offset to seed the in-sync offset of the STM (off by
one). Because of that the STM may skip the first record batch (which is
usually a raft_configuration batch which doesn't affect the archival
STM).

Also, the last applied offset of the snapshot is seeded incorrectly
using the last applied offset of the manifest. The commit fixes this by
using last uploaded offset.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>