datalake: enforce disk usage in serde parquet writer
This commit adds a disk writer tracker interface, analgous to the mem
tracker interface, which is used to track disk usage of translators and
enforce a limit on usage based on reservations.

When a translator's reservation is exhausted it requests more units from
the scheduler's resource pool, introduced in the previous commit. If it
is not granted units then it halts execution waiting on units to arrive
from the global pool.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>