`storage`: set `index_state`'s `base_offset` during compaction
This value would be incorrectly persisted as `0` during and after
compaction.

This means batches indexed in the `index_state` during compaction
would have an incorrect `relative_offset` added, and furthermore,
the value of `base_offset` would forever be `0` on disk.

Fix this by adding `base_offset` as a required parameter in
`index_state::make_empty_index()` and setting the `index_base_offset`
when constructing the `copy_data_segment_reducer` for use with self
compaction and sliding window compaction.