build-ducktape-image: use download cache
Use a --mount=type=cache to mount a cache directory for large downloads.
This is based on the observation that layers are rebuild much more
frequently than the downloads are updated. This should speed up image
rebuilds significantly, especially when the localhost is far from the
download servers.

The README.md is updated and contains additional details on this change.

After this change the image build time is reduced from 15-60m (network
dependent) to 7-8m. The critical path after this change is through
nessie which has a long uncached gradle download/build.