c/partition_balancer: track detailed information about failed reallocations
When partition replicas area moved in the cluster from whichever reason
Redpanda should expose all information required to identify issues
experienced during that operation. Previously partition balancer planner
included only the basic information about the failures in the status.
That was the list of `ntps` that replicas failed to be reallocated and
the total number of reallocations failures. In order to obtain more
detailed information looking into the partition balancer log was
required. Moreover the reallocations that were not even attempted during
preparing an allocation plan were not reported at all. This lead to
confusion as f.e. partitions with only one replica on the node that is
offline that can not be moved was not reported ans the failed one.

In order to address this problem we introduced a new
`allocation_failure_details` map in the partition balancer planner
result. The map provides a detailed information about the failed
reallocation. Additionally reallocation failure is reported when
partition is marked as `immutable` in the process of planning balancer
action. This makes it easy for the operator to identify any partition
balancer issues without the need to look for the specific log entries.

The change is fully backward compatible as all the changes are
incremental i.e. when no detailed information about failures is
available Redpanda will fallback to the old basic `ntp` list.

Signed-off-by: Michał Maślanka <michal@redpanda.com>