`storage`: add "deprecated" code comments to `compacted_index`
In the `compacted_index`, we now only ever write entries of `entry_type::key`
via the `spill_key_index`. Note this in the `compacted_index::entry_type` enum
for future reference.

Also, the `compacted_index::footer_flag::truncation` bitflag is not used
in the tree either. Append a "deprecated" comment here as well.