kafka: Add cloud_topic_partition
This commit adds a 'cloud_topic_partition' class which is an
implementation of the 'partiton_proxy'. The implementation connects the
control plane (currently just cluster::partition) with the data plane
(cloud_topics::data_plane).

The main goal here is to separate the control plane and data plane.
There is no explicit API boundary for the control plane yet. The
'cluster::partition' is queried directly. But it will be possible to
introduce the explicit control plane API in the future.

The data plane API boundary is defined by the
'cloud_topics::data_plane_api'.

The 'cloud_topic_partition' instance queries 'dl_placeholder' batches
and converts them to 'extent_meta' type in the read path. In the write
path it sends user data to the data plane (write_and_debounce method)
and gets 'extent_meta' struct back. Then it coverts this 'extent_meta'
struct to 'dl_placeholder' batch and replicates it using the underlying
partition.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>