container: Introduce chunked_circular_buffer
Just an alias for std::deque for now, but this will make it easier to swap out
for a custom container if needed.

Of particular concern is the lack of control over chunk size. Quick googling
suggests that, in clang, it should be the larger of 4KiB. We could build
something similar on top of fragmented_vector if we want more control.