`storage`: rework adjacent segment compaction with ranges
Removes the now obselete `log_compaction_adjacent_merge_self_compaction_count`
tunable, and removes the functions `find_adjacent_compaction_range()` and
`compact_adjacent_segments()` from `disk_log_impl`.

Adjacent segment compaction is now completely replaced with the implementation
for ranges of segments, allowing us to compact more than two adjacent segments
at a time.

The routine is similar to sliding window compaction, in that all segments in
the set are iterated over and self compacted before we move on to the "meat"
of the routine (in this case, the actual adjacent merging).