`storage`: add `self_compact_timestamp` to `index_state`
To mark the timestamp at which a `segment` is first self compacted.

This is going to be eventually needed for the removal of transactional
control batches using `delete.retention.ms` (the time horizon for removal of
transactional control batches is `self_compact_timestamp + delete.retention.ms`,
analogous to the time horizon for tombstone records set by `clean_compact_timestamp
+ delete.retention.ms`).

The reasons as to why the `self_compact_timestamp` is meaningful for
transactional batch removal are:

1. We already remove aborted transactional data during self compaction.
2. We are guaranteed in the current compaction implementation to self
   compact segments from oldest to newest- so even transactions which
   span segments (which is already an unlikely case) will correctly
   follow the two pass approach of:
   i. removing the beginning tx fence batch, unsetting the transactional
      bit within all Raft data batches associated with the transaction,
      and setting `self_compact_timestamp`
   ii. removing the control batch after `self_compact_timestamp + delete.retention.ms`

in that order.

To summarize, the condition for setting the base timestamp that defines the
transactional batch removal horizon is analagous to tombstones. That is, the
base timestamp for tombstone removal can only be set when every record sharing
the tombstone's key is removed by compaction, just like the base timestamp for
transactional batch removal can only be set when the `tx_fence` batch has been
removed _and_ all Raft data batches have had their transactional bit unset.

`IndexState.BinaryCompatibilityTest` must now be removed (it was intended to be
an ephemeral test anyways).