cloud_storage_clients: do not wait when not needed
It may happen that local pool got back some clients in between checking
that it is empty and failing to acquire a remote client.

In the worst case scenario this may lead to temporary underutilization
of the client pool. It is highly improbable to happen in real case.

No test as it would require pretty involved refactoring of the client
pool to introduce failure injection (delays) for cross shard calls.