cloud_storage_clients: destroy oldest client rather than newest on borrow
The _pool variable holds clients in such a way that the latest used
client is always at the top of the stack. We push_back() during release.

The client at the top of the stack (given by back() call) is likely to
have a fresher connection (non-expired). Since we are destroying a
client, destroy from the bottom of the stack (given by front() call)
which is more likely to have an expired connection.