k/client: Added support to fetch metadata
Added method to fetch metadata from a source cluster and return the
results.  Added unit test to ensure that metadata response contains
cluster and topic authorized operations.  If Redpanda updates which
version of Metadata it supports to a version that no longer supplies
those responses, the test will fail, indicating that DescribeCluster is
being supported.

Signed-off-by: Michael Boquard <michael@redpanda.com>