datalake: fix iceberg primitive value translation
Before this commit redpanda would fail to translate Avro schemas which
contained a primitive type at the schema root. Trying to translate such
a schema failed with the following exception:

```
Avro schema translation failed: avro::Exception (This type does not have
attribute)
```

It appears that Avro primitive types do not support names but in the
code we had a call for getting that name.

The other observation is that we never used that type name so I removed
it altogether instead of adding conditionals depending on whether the
root schema has a name or not.