dt/e2e_si: Traffic shaping test
General idea here is to trigger timeouts in 'cloud_io::remote' by adding
delays to traffic from the redpanda node -> s3. To that end, add a stub
node interface that wraps the IP address of the s3 service and use
ClusterTopology as usual.

Note that we have no way to shape the other direction of the link (s3 -> RP)

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>