`storage`: add `segment::set_cached_disk_usage()`
Sets the cached disk usage for the `segment` and `compacted_index`. These
are usually the values from the underlying `segment_{appender/reader}`
for either of these objects. The base `segment_index` sets its disk usage
in `flush_to_file()`.

This function is used within the compaction subsystem to cut down on
future calls to `stat()`, since the cached disk usage is cleared during
every round of compaction, and we know from in-memory state how large
these objects are going to be on disk.