`storage`: co-routinize `copy_filtered_entries()`
Exception handling should remain the same here in co-routinized version
due to `try/catch` blocks and capturing of exception with `eptr`.

There is one slight functional change here, which is that `writer->set_flag()`
is ONLY called if `reader.consume()` does not throw (which feels more correct
than setting the flag even in the presence of a thrown exception).