r/tests: further increase tolerance in leadership_transfer_delay test
The test validates if the election after reconfiguration is taking
roughly the same amount of time as the one when transferring
leadership. The problem with too tight tolerance causes the test to be
flaky. Increased the tolerance a little bit as even with 70% of
tolerance the value is still much smaller than the election timeout.

Signed-off-by: Michał Maślanka <michal@redpanda.com>