datalake: add default partition spec kludge for AWS Glue
AWS Glue does not support partitioning on nested fields, which means it
does not support partitioning on the default `(hour(redpanda.timestamp))`
partition spec. In fact, trying to use a partition spec on a nested
fields causes 500 errors from Glue on catalog operations. To work around
this, default the partition spec to empty `()` when we detect we're
using AWS Glue as the Iceberg REST catalog.