rptests/tests: Disable retention period in consumer group stress test
Sometimes, this test would take more than 10 minutes to run, which
is the default value of `group_offset_retention_check_ms`. This would
trigger the `group_offset_retention_sec`, which was set to only 20sec
in this test. Then when the client tries to get the offsets, after
the consumers have been stopped, it would return an error value (-1001).

As this test is purpose is correctness under heavy load, not performance,
offset retention is disabled.