ct/l1: index partitions by ntp in footer
A few changes, first of all we rename `object_index` to footer. Then we
use a multimap from ntp to the file positions and index entries for each ntp.

In addition to this, there seems to be some issues in serde with
non-copyable map values, so we don't use serde_fields but instead
specify how to write these values manually.