tests: add infra for analyzing shutdown hangs
In https://github.com/redpanda-data/redpanda/pull/25594 there was some
structured logging added to help diangose shutdown hangs. It logs
messages about what is being shutdown and what is next ot shutdown.

This commit adds some very basic infrastructure for searching logs and
analyzing these logs in response to a redpanda service shutdown that
fails due to timeout.

It looks for the last service expected to shutdown and then looks for
the message indicating that this last service was shutdown. If it isn't
found, then it indicates that perhaps that service is what hung.

[DEBUG - 2025-06-29 20:49:13,706 - redpanda - _log_node_shutdown_analysis - lineno:3462]: Last structured shutdown line: INFO  2025-06-29 20:49:13,258 [shard 0:main] main - application.h:307 - Stopping resources::cpu_profiler, ..next to shutdown is memory_sampling
[DEBUG - 2025-06-29 20:49:13,706 - redpanda - _log_node_shutdown_analysis - lineno:3467]: Found stopping message for last service INFO  2025-06-29 20:49:13,258 [shard 0:main] main - application.h:309 - Stopping memory_sampling

This isn't perfect by any means but can help avoid some pain in looking
through logs.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>