tests: dt matrix must be json de/en-codable
In a test execution without parametrizing the tests, the matrix can
return any type that's compatible with python. However, when a test is
being parametrized, ducktape expects a json string. Having tuples work
with python but ducktape cannot json-decode it leading ducktape to not
be able to load that test. For instance, the test_tiered_storage test
calls a function to parametrize the matrix and that must return a json
decodeable output. This was blocking this test from being retried with
the error:

```
[WARNING:2025-06-06 23:26:21,518]: No tests loaded for /root/tests/rptest/tests - tiered_storage_model_test.py - TieredStorageTest - test_tiered_storage - {'cloud_storage_type_and_url_style': [1, 'path'], 'test_case': {'name': '(TS_Read == True, TS_TxRangeMaterialized == True, SpilloverManifestUploaded == True)'}}
[WARNING:2025-06-06 23:26:21,518]: Didn't find any tests in /root/tests/rptest/tests/tiered_storage_model_test.py
```

jira: [DEVPROD-3062]