storage: quiet logging when no tiered storage topics
Space management can be turned on when TS is not enabled, which won't do
anything but we still end up generating a lot of log messages. One
option would be to only log or run the workload when TS is enabled, but
the log messages being complained about would also be logged if TS were
enabled and no TS topics had yet been created (e.g. in a fresh cluster).

So in this commit we log at debug level if it appears there isn't any TS
topics in the system which should cover both cases.

Fixes: https://github.com/redpanda-data/redpanda/issues/24505

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>