rptest: speed up cluster metadata listing
The cluster_metadata_upload_test waits 30 seconds for a bucket listing
to show valid cluster metadata. On ABS (empirically slow in CI), the
presence of several partition manifests delays the discovery of cluster
metadata for that entire duration. If this happens after the metadata
uploader has removed some stale cluster metadata already, the test fails
to download/validate the cluster metadata snapshot on its first and only
try, and the entire test fails.

This updates the test to filter the cluster metadata listing by
cluster_metadata to speed the listing up.