test_utils: continue running tests after exceptions
Previously gtest would fail fatally (with reports from LSAN even) when
we would use our gtest_main implementation. We don't want that. Instead,
we want to fail only a single test and continue with the rest.

This in part is caused by the code we added in 6f707f5c22 to trigger
fatal failures on failed asserts from test helpers and some gtest
internals.

When throwing an exception, gtest handles it and reports a fatal test
result but then we throw again which aborts entire test run as it is not
prepared to handle such exceptions.

Fix the problem by avoiding throwing on test failures caused by
exceptions. The check is not ideal but it does work and a test is
included.

We should propose a proper mechanism upstream for handling this problem
in the long run.