wcr: skip creation of all internal topics
Before this commit during `get_actions depending on topic properties we
would either put a topic in `actions.remote_topics` or
`actions.local_topics`. Later, if we detected that a topic is internal
during processing of remote topics we would skip creating that topic.

To best of my knowledge there is only one topic that could fall into
this category, `__consumer_offsets`.

The result is that post-recovery there is no automatic creation of
`__consumer_offsets` topic.

While I was experimenting with disabling TS for consumer offsets topic I
noticed that topic now gets automatically created during recovery.
Although it is likely that this won't be the actual way we'll disable TS
for consumer offsets I found that behavior surprising enough to prompt
me to move the check earlier in the code such that we never attempt to
create internal topics during recovery.

I believe this makes it more straightforward to understand what topics
are created and not created during WCR rather than before.

We also have clusters in the wild with manually disabled TS for CO
topic. When/if these get recovered they would get the same surprising
behavior.

This fixes no real bugs and shouldn't introduce any new bugs.