c/a: Introduce archival_metadata_stm::cloud_recoverable_offset
Returns the maximum offset which is guaranteed to be recoverable from
cloud storage.

This is the lesser of the last offset uploaded to cloud storage and the
last offset we uploaded a manifest for.

If the manifest is empty or the last uploaded offset is 0, returns
offset::min(), indicating that no data is recoverable from cloud.