c/a/test: Check that term doesn't change across sync call
These tests are intended to check that sync operations w/in a single
raft term behave in a certain way. Previously the test would assert that
the rafter term after a successful sync operation matches a specific,
hard-coded value.

This seems to be a fine assumption in the vast majority of cases, but the test
has failed at least once in CI. It appears that this assertion could fail
in either of two cases:
1. sync unexpectedly succeeded across a raft term boundary, in which case the
   test should fail.
2. leader election took place _prior_ to kicking off the sync operation, in
   which case we should be asserting on the term ID after that election and the
   test should pass.

This commit changes the assert to instead compare the term after sync to the
term before sync. This should eliminate the possibility of asserting on (2).

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>