iceberg: serialize Struct fields as JSON string columns
This adds support for protobuf Struct, Value, and ListValue well-known
recursive types. Instead of creating nested fields, it makes a string
column and serializes the value as a JSON string. This is a simple way
to handle such fields and has been requested by a number of users.
Many query engines can deserialize and query the JSON string, so it's
pretty useful.