`raft`: add `_remake_cb` and `set/clear` functions to `group_manager`
The request to remake a `partition` will bubble up from `raft` into
the `cluster` layer, rather than a direct line of communication from an
intermediate object which depends on both `raft` and `cluster` libraries.

The idea is that the `consensus` will invoke a callback registered by
the `group_manager`, which in turn will invoke a callback registered by
the `controller` between the `group_manager` and `controller_api` to
invoke a `remake_partition()` command.

This commit adds the `_remake_cb` as well as the `set/clear_remake_cb()`
functions. Since the lifetime of the `group_manager` and `controller_backend/api`
are not tied, `clear_remake_cb()` needs to be used to ensure no calls to
stopped services are made (in the situation that we receive a `remake_partition()`
request in the middle of shutdown).