bazel/packaging: allow interpreter path to be configured by parameter
The interpreter path in elf binary can not be relative therefore the
interpreter path depends on the binary install path. In our current
setup with the ducktape tests Redpanda binary install path in ducktape
container depends on the Redpanda source tree location in local
filesystem. In order to make using the tarball or directory packages
easier and not requiring the script to override the interpreter made the
interpreter path configurable. This when starting tests the package will
be built with correct interpreter path set

Signed-off-by: Michał Maślanka <michal@redpanda.com>