c/a: Re-foreground segment uploads
8f84b6d01ce40b68196a1584d91db999fc1a6372 pushed segment index uploads
into the background, which should in principle free the upload loop
for additional work as index uploads are not strictly required for
correctness and metadata replication can proceed before they complete.

However, we have some suspicion that there could be a negative impact
associated with backgrounding these requests in some edge cases, so
here we put the index upload back in the foreground pending further
investigation.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>