ducktape: upgrade to python 3.11
This is a rather lot of change, although the diff is small.

First all of instead of using distro versions or the deadsnakes PPA,
instead download a standalone build of python (used by many projects)
so we have complete control over the exact version of python and can
upgrade at will.

Secondly, we upgrade ducktape to point to updated deps. Previously it
pointed to a forked version of boto3 with FIPS support, but FIPS is now
supported in the upstream boto3/botocore releases. So drop our fork for
a more recent version.

Third we switch to a custom published open polaris. They no longer check
in their generated client, and it's also not published anywhere. So
instead we have a new repo that generates the client and publishes the
client to PyPI. That package is consumed directly here.

Lastly, upgrade a few other dependencies that didn't have support for
the latest python version we're using, those are simple version bumps.